refactor: grouping component files
