+++ src/components/Card/Card.scss
@@ -2,6 +2,7 @@
     display: flex;
     flex-direction: column;
     row-gap: 32px;
+    color: var(--tds)
 }
 
 .split-cards {
