+++ src/components/Modal/Modal.scss
