+++ src/components/Modal/Modal.tsx
@@ -8,23 +8,7 @@ type ModalProps = {
 
 const Modal = ({ selector, size, prevent }: ModalProps) => {
   return (
-    <tds-modal
-      ref={(element) => {
-        /* if (size === "lg") {
-          element?.addEventListener("tdsClose", (event) => {
-            event.preventDefault();
-            alert(
-              `This Modal prevent default of the tdsClose event, so it won't close the way it normally does. You can close it by using the Modals closeModal() function.`
-            );
-          });
-        } */
-      }}
-      prevent={prevent}
-      selector={selector}
-      id="my-modal"
-      size={size}
-      actions="static"
-    >
+    <tds-modal prevent={prevent} selector={selector} id="my-modal" size={size} actions="static">
       <h5 className="tds-modal-headline" slot="header">
         Modal
       </h5>
