+++ src/components/Stepper/Stepper.scss
