+++ src/pages/StepperPage.tsx
@@ -1,12 +1,12 @@
-import Stepper from '../components/Stepper';
+import Stepper from '../components/Stepper/Stepper';
 
 const StepperPage = () => {
-    return (
-        <div>
-            <Stepper labelPosition="below" orientation="vertical" />
-            <Stepper labelPosition="aside" orientation="horizontal" />
-        </div>
-    );
+  return (
+    <div>
+      <Stepper labelPosition="below" orientation="vertical" />
+      <Stepper labelPosition="aside" orientation="horizontal" />
+    </div>
+  );
 };
 
 export default StepperPage;
