+++ src/pages/WebComponents.tsx
@@ -1,13 +1,13 @@
 /* eslint-disable jsx-a11y/anchor-is-valid */
 import './WebComponents.scss';
 import Banner from '../components/Banner';
-import Card from '../components/Card';
+import Card from '../components/Card/Card';
 import Button from '../components/Button';
-import Toast from '../components/Toast';
-import Modal from '../components/Modal';
-import Toggle from '../components/Toggle';
+import Toast from '../components/Toast/Toast';
+import Modal from '../components/Modal/Modal';
+import Toggle from '../components/Toggle/Toggle';
 import Block from '../components/Block';
-import RadioButton from '../components/RadioButton';
+import RadioButton from '../components/RadioButton/RadioButton';
 import { PopoverMenu } from '../components/PopOverCanvas';
 import PopoverCanvas from '../components/PopOverCanvas';
 import AccordionExample from '../components/Accordion';
@@ -86,7 +86,9 @@ export default function WebComponents() {
         open-direction="auto"
       >
         <tds-dropdown-option value="option-1">Option 1</tds-dropdown-option>
-        <tds-dropdown-option disabled value="option-2">Option 2</tds-dropdown-option>
+        <tds-dropdown-option disabled value="option-2">
+          Option 2
+        </tds-dropdown-option>
         <tds-dropdown-option value="option-3">Option 3</tds-dropdown-option>
         <tds-dropdown-option value="option-4">Option 4</tds-dropdown-option>
         <tds-dropdown-option value="option-5">Option 5</tds-dropdown-option>
